+++ .storybook/preview.ts
@@ -1,4 +1,5 @@
 import { Preview } from "@storybook/react";
+import { configure } from "@storybook/test";
 
 import "../src/style/fonts.css";
 
@@ -8,6 +9,9 @@ import withGlobalStyles from "./with-global-styles";
 import withLocaleSelector from "./with-locale-selector";
 import withPortalProvider from "./with-portal-provider";
 import { withThemeProvider, globalThemeProvider } from "./withThemeProvider";
+
+// Configure the testIdAttribute to look for data-role when querying elements using `getByTestId`.
+configure({ testIdAttribute: "data-role" });
 
 const customViewports = {
   xsm: { name: "Extra small", styles: { width: "320px", height: "900px" } },
@@ -36,7 +40,7 @@ const parameters = {
   },
   chromatic: { disableSnapshot: false },
   viewport: { viewports: customViewports },
-  viewMode: process.env.STORYBOOK_VIEW_MODE,
+  viewMode: import.meta.env.STORYBOOK_VIEW_MODE,
 };
 
 const globalTypes = {