+++ src/components/action-popover/action-popover-item/action-popover-item.component.tsx
@@ -6,7 +6,6 @@ import {
   SubMenuItemIcon,
   StyledMenuItem,
   StyledMenuItemInnerText,
-  StyledMenuItemOuterContainer,
   StyledMenuItemWrapper,
 } from "../action-popover.style";
 import Events from "../../../__internal__/utils/helpers/events";
@@ -41,25 +40,11 @@ export interface ActionPopoverItemProps {
   /** Submenu component for item */
   submenu?: React.ReactNode;
   /** @ignore @private */
-  placement?: "top" | "bottom";
-  /** @ignore @private */
   focusItem?: boolean;
-  /** @ignore @private */
-  horizontalAlignment?: Alignment;
-  /** @ignore @private */
-  childHasSubmenu?: boolean;
-  /** @ignore @private */
-  childHasIcon?: boolean;
   /** @ignore @private */
   currentSubmenuPosition?: Alignment;
   /** @ignore @private */
-  setChildHasSubmenu?: (value: boolean) => void;
-  /** @ignore @private */
-  setChildHasIcon?: (value: boolean) => void;
-  /** @ignore @private */
   setCurrentSubmenuPosition?: (value: Alignment) => void;
-  /** @ignore @private */
-  isASubmenu?: boolean;
 }
 
 const INTERVAL = 150;
@@ -102,18 +87,11 @@ export const ActionPopoverItem = ({
   disabled = false,
   onClick: onClickProp,
   submenu,
-  placement = "bottom",
   focusItem,
   download,
   href,
-  horizontalAlignment,
-  childHasSubmenu,
-  childHasIcon,
   currentSubmenuPosition,
-  setChildHasSubmenu,
-  setChildHasIcon,
   setCurrentSubmenuPosition,
-  isASubmenu = false,
   ...rest
 }: ActionPopoverItemProps) => {
   invariant(
@@ -121,7 +99,7 @@ export const ActionPopoverItem = ({
     "ActionPopoverItem only accepts submenu of type `ActionPopoverMenu`",
   );
 
-  const { setOpenPopover, isOpenPopover, focusButton, submenuPosition } =
+  const { setOpenPopover, focusButton, submenuPosition } =
     useActionPopoverContext();
   const isHref = !!href;
   const [containerPosition, setContainerPosition] = useState<
@@ -135,22 +113,6 @@ export const ActionPopoverItem = ({
   const ref = useRef<HTMLButtonElement>(null);
   const mouseEnterTimer = useRef<NodeJS.Timeout | null>(null);
   const mouseLeaveTimer = useRef<NodeJS.Timeout | null>(null);
-
-  useEffect(() => {
-    /* istanbul ignore if - doesn't seem to actually run as the child item is unmounted before the context updates */
-    if (!isOpenPopover) {
-      setOpen(false);
-    }
-  }, [isOpenPopover]);
-
-  useEffect(() => {
-    if (icon) {
-      setChildHasIcon?.(true);
-    }
-    if (submenu) {
-      setChildHasSubmenu?.(true);
-    }
-  }, [icon, setChildHasSubmenu, setChildHasIcon, submenu]);
 
   const alignSubmenu = useCallback(() => {
     const checkCalculatedSubmenuPosition = calculateSubmenuPosition(
@@ -180,19 +142,14 @@ export const ActionPopoverItem = ({
       const leftAlignedSubmenu = currentSubmenuPosition === "left";
       const leftValue = leftAlignedSubmenu ? -submenuWidth : "auto";
       const rightValue = leftAlignedSubmenu ? "auto" : -submenuWidth;
-      const yPositionName =
-        placement === "top"
-          ? /* istanbul ignore next - tested in Playwright */ "bottom"
-          : "top";
 
       return {
         left: leftValue,
-        [yPositionName]: "calc(-1 * var(--spacing100))",
         right: rightValue,
       };
     };
     setContainerPosition(getContainerPosition);
-  }, [submenu, currentSubmenuPosition, placement]);
+  }, [submenu, currentSubmenuPosition]);
 
   useEffect(() => {
     if (submenu) {
@@ -330,90 +287,59 @@ export const ActionPopoverItem = ({
     }),
   };
 
-  const renderMenuItemIcon = () => {
-    return (
-      icon && (
-        <MenuItemIcon
-          type={icon}
-          data-element="action-popover-menu-item-icon"
-          horizontalAlignment={horizontalAlignment}
-          submenuPosition={currentSubmenuPosition}
-          childHasIcon={childHasIcon}
-          childHasSubmenu={childHasSubmenu}
-          hasIcon={!!icon}
-          hasSubmenu={!!submenu}
-          isASubmenu={isASubmenu}
-        />
-      )
-    );
-  };
-
   return (
-    <StyledMenuItemWrapper {...(submenu && wrapperProps)}>
-      <div onKeyDown={onKeyDown} role="presentation">
-        <StyledMenuItem
-          {...rest}
-          ref={ref}
-          onClick={onClick}
-          type="button"
-          tabIndex={0}
-          isDisabled={disabled}
-          horizontalAlignment={horizontalAlignment}
-          submenuPosition={currentSubmenuPosition}
-          hasSubmenu={!!submenu}
-          childHasSubmenu={childHasSubmenu}
-          {...(disabled && { "aria-disabled": true })}
-          {...(isHref && { as: "a" as unknown as undefined, download, href })}
-          {...(submenu && itemSubmenuProps)}
-        >
-          {submenu && checkRef(ref) && currentSubmenuPosition === "left" ? (
-            <SubMenuItemIcon
-              data-element="action-popover-menu-item-chevron"
-              type="chevron_left_thick"
-            />
-          ) : null}
-          <StyledMenuItemOuterContainer>
-            {horizontalAlignment === "left" ? renderMenuItemIcon() : null}
-            <StyledMenuItemInnerText
-              data-element="action-popover-menu-item-inner-text"
-              horizontalAlignment={horizontalAlignment}
-              submenuPosition={currentSubmenuPosition}
-              isASubmenu={isASubmenu}
-              childHasSubmenu={childHasSubmenu}
-              childHasIcon={childHasIcon}
-              hasIcon={!!icon}
-              hasSubmenu={!!submenu}
-            >
-              {children}
-            </StyledMenuItemInnerText>
-            {horizontalAlignment === "right" ? renderMenuItemIcon() : null}
-          </StyledMenuItemOuterContainer>
-          {submenu && checkRef(ref) && currentSubmenuPosition === "right" ? (
-            <SubMenuItemIcon
-              data-element="action-popover-menu-item-chevron"
-              type="chevron_right_thick"
-            />
-          ) : null}
-        </StyledMenuItem>
-        {React.isValidElement(submenu)
-          ? React.cloneElement<ActionPopoverMenuProps>(
-              submenu as React.ReactElement<ActionPopoverMenuProps>,
-              {
-                parentID: `ActionPopoverItem_${guid}`,
-                menuID: `ActionPopoverMenu_${guid}`,
-                "data-element": "action-popover-submenu",
-                isOpen,
-                ref: submenuRef,
-                style: containerPosition,
-                setOpen,
-                setFocusIndex,
-                focusIndex,
-                isASubmenu: true,
-                horizontalAlignment,
-              },
-            )
-          : null}
-      </div>
+    <StyledMenuItemWrapper onKeyDown={onKeyDown} {...(submenu && wrapperProps)}>
+      <StyledMenuItem
+        {...rest}
+        ref={ref}
+        onClick={onClick}
+        type="button"
+        tabIndex={0}
+        isDisabled={disabled}
+        {...(disabled && { "aria-disabled": true })}
+        {...(isHref && { as: "a" as unknown as undefined, download, href })}
+        {...(submenu && itemSubmenuProps)}
+      >
+        {submenu && checkRef(ref) && (
+          <SubMenuItemIcon
+            aria-hidden
+            data-element="action-popover-menu-item-chevron"
+            data-role="chevron-icon"
+            type={
+              currentSubmenuPosition === "left"
+                ? "chevron_left_thick"
+                : "chevron_right_thick"
+            }
+          />
+        )}
+        {icon && (
+          <MenuItemIcon
+            aria-hidden
+            type={icon}
+            data-element="action-popover-menu-item-icon"
+            data-role="item-icon"
+          />
+        )}
+        <StyledMenuItemInnerText data-element="action-popover-menu-item-inner-text">
+          {children}
+        </StyledMenuItemInnerText>
+      </StyledMenuItem>
+      {React.isValidElement(submenu)
+        ? React.cloneElement<ActionPopoverMenuProps>(
+            submenu as React.ReactElement<ActionPopoverMenuProps>,
+            {
+              parentID: `ActionPopoverItem_${guid}`,
+              menuID: `ActionPopoverMenu_${guid}`,
+              "data-element": "action-popover-submenu",
+              isOpen,
+              ref: submenuRef,
+              style: containerPosition,
+              setOpen,
+              setFocusIndex,
+              focusIndex,
+            },
+          )
+        : null}
     </StyledMenuItemWrapper>
   );
 };