+++ src/components/date-range/index.ts
@@ -1,3 +1,6 @@
 export { default } from "./date-range.component";
-export type { DateRangeChangeEvent } from "./date-range.component";
+export type {
+  DateRangeChangeEvent,
+  DateRangeProps,
+} from "./date-range.component";
 export type { DateRangeContextProps } from "./__internal__/date-range.context";